test(ModalFooter): move test for ModalFooter component on a separate test file
