+++ src/components/feedback/Modal.test.tsx
@@ -10,7 +10,6 @@ import { screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
 import { Modal, ModalProps } from './Modal';
-import { ModalFooter } from './modal-components/ModalFooter';
 import { render } from '../../test-utils';
 import { Button } from '../basic/Button';
 import { Text } from '../basic/Text';
@@ -105,25 +104,5 @@ describe('Modal', () => {
 			expect.stringContaining('Error: Not implemented: navigation (except hash changes)')
 		]);
 		console.error = originalConsoleError;
-	});
-
-	describe('Modal footer', () => {
-		it('displays a disabled primary button if the "confirmDisabled" is set to true', async () => {
-			render(<ModalFooter confirmLabel={'confirm'} confirmDisabled={true} onConfirm={jest.fn} />);
-			const confirmButton = screen.getByRole('button', { name: /confirm/i });
-			expect(confirmButton).toBeDisabled();
-		});
-
-		it('displays an enabled primary button if the "confirmDisabled" is set to false', async () => {
-			render(<ModalFooter confirmLabel={'confirm'} confirmDisabled={false} onConfirm={jest.fn} />);
-			const confirmButton = screen.getByRole('button', { name: /confirm/i });
-			expect(confirmButton).toBeEnabled();
-		});
-
-		it('displays an enabled primary button if the "confirmDisabled" is not set', async () => {
-			render(<ModalFooter confirmLabel={'confirm'} onConfirm={jest.fn} />);
-			const confirmButton = screen.getByRole('button', { name: /confirm/i });
-			expect(confirmButton).toBeEnabled();
-		});
 	});
 });
